download-openfoodfacts-products-jsonl/requirements.txt
requests==2.28.1
google-cloud-storage==2.5.0